+++ backend/batan/src/lib.rs
@@ -1,6 +1,4 @@
 pub mod consumer;
 pub mod job;
 pub mod producer;
-pub mod queue;
-pub mod retry;
 pub mod tracer;
